New updates

+++ database/migrations/2022_06_22_094730_create_categories_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('slug')->unique()->default('yes');
             $table->text('description')->nullable();
-            $table->unsignedInteger('parent_id')->default(1)->nullable();
+            $table->unsignedBigInteger('parent_id')->default(1)->nullable();
             $table->boolean('featured')->default(0);
             $table->boolean('menu')->default(1);
             $table->string('image')->nullable();
